Correctly implemented NEAP and Time of Arrival.  Fixed case incidence eqn 2.

+++ src/sample/Route/MultiPath.java
@@ -41,9 +41,9 @@ public class MultiPath extends Route {
 
 
     @Override
-    public int getGumbelPrediction()
+    public double getGumbelPrediction()
     {
-        int pred = 0;
+        double pred = 0;
         for(Path path : pathList)
             pred += path.getGumbelPrediction();
