+++ src/sample/Route/Route.java
@@ -6,7 +6,7 @@ import sample.Disease;
 public abstract class Route {
     public abstract double getNEAP();
     public abstract double getEffDis();
-    public abstract int getGumbelPrediction();
+    public abstract double getGumbelPrediction();
     public abstract CityNode getSourceCity();
     public abstract CityNode getTargetCity();
 }
